Credential check with session creation
In lieu of checking the credentials separately, either by opening the
cred file and parsing it or by invoking an API call to a separate
microservice, SESSION CREATE now requires the user's name and
authentication key, and only returns a session if the credentials are
valid.

Session manager will handle credential management as well as sessions.

Add SESSION keyword layer to session manager microservice interface.